--- src/components/common/BottomSheet/ButtonGroup/ButtonGroup.jsx
@@ -1,5 +1,0 @@
-import { Children } from 'react';
-
-export default function ButtonGroup() {
-  return <div>{Children}</div>;
-}
