--- src/components/common/BottomSheet/ButtonGroup/ButtonGroup.module.css
@@ -1,5 +1,0 @@
-.group {
-  display: flex;
-  gap: 0.5rem;
-  width: 100%;
-}
