+++ src/components/common/BottomSheet/ClassCreateForm.jsx
@@ -1,12 +1,12 @@
-import styles from './BottomSheetForm.module.css';
-import Button from '../../Button/Button';
-import Input from '../../Input/Input';
-import { useBottomSheet } from '../../../../context/BottomSheetContext';
-import useInput from '../../../../hooks/useInput';
-import CheckBox from '../../Input/CheckBox';
-import useClassList from '../../../../hooks/useClassList';
-import ModalPortal from '../../Modal/ModalPortal';
-import ToastNotification from '../../Modal/ToastNotification';
+import styles from './css/BottomSheetForm.module.css';
+import Button from '../Button/Button';
+import Input from '../Input/Input';
+import { useBottomSheet } from '../../../context/BottomSheetContext';
+import useInput from '../../../hooks/useInput';
+import CheckBox from '../Input/CheckBox';
+import useClassList from '../../../hooks/useClassList';
+import ModalPortal from '../Modal/ModalPortal';
+import ToastNotification from '../Modal/ToastNotification';
 
 export default function ClassCreateForm({ tag }) {
   const { toggle } = useBottomSheet();
