--- src/components/common/BottomSheet/Header/Header.jsx
@@ -1,5 +1,0 @@
-import styles from './Header.module.css';
-
-export default function Header({ children }) {
-  return <header className={styles.header}>{children}</header>;
-}
