+++ src/components/common/BottomSheet/SheetHeader.jsx
@@ -1,0 +1,5 @@
+import styles from './css/SheetHeader.module.css';
+
+export default function SheetHeader({ children }) {
+  return <header className={styles.header}>{children}</header>;
+}
