+++ src/pages/Detail/Detail.jsx
@@ -7,7 +7,7 @@ import styles from './Detail.module.css';
 import LoadingDetail from '../../components/common/LoadingDetail/LoadingDetail';
 import BottomSheet from '../../components/common/BottomSheet/BottomSheet';
 import { DEPOSIT, WITHDRAW } from '../../constants/bottomSheetTag';
-import Body from '../../components/common/BottomSheet/Body/Body';
+import SheetBody from '../../components/common/BottomSheet/SheetBody';
 import HistoryList from './HistoryList';
 import useDetail from './hooks/useDetail';
 import useClassDetail from '../../hooks/useClassDetail';
@@ -39,14 +39,14 @@ export default function Detail() {
             <BottomSheet.Button text='출금' type='button' tag={WITHDRAW} />
           </BottomSheet.Header>
           <BottomSheet.Body>
-            <Body.TransferForm
+            <SheetBody.TransferForm
               code={code}
               histories={detail.history}
               text='입금'
               tag={DEPOSIT}
               color='red'
             />
-            <Body.TransferForm
+            <SheetBody.TransferForm
               code={code}
               histories={detail.history}
               text='출금'
